pyproject.toml
[project]
name = "langStuff"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datetime>=5.5",
    "fastapi>=0.116.1",
    "grandalf>=0.8",
    "ipdb>=0.13.13",
    "ipython>=9.4.0",
    "json5>=0.12.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "paperclip>=2.7.4",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
    "uuid>=1.30",
    "uvicorn[standard]>=0.35.0",
]
